Implement a new system for managing cradle files in testdata.
We want to test HLS. So we need to make up some fake Haskell code to test HLS
on. And we need to make up fake cradles so HLS can load our fake code. But while
are developing, the real HLS sees the fake cradles and freaks out. So we need to
hide the fake cradles until its time to run the test suite.

We can write the cradles we need, but we call them "test-hie.yaml" instead of
"hie.yaml", that way they don't get detected as cradle files until we are ready
for them to be used. In the test suite, we bracket our tests by copying
"test-hie.yaml" to "hie.yaml" and deleting the copy.